nautilus_backtest::exchange

Struct SimulatedExchange

pub struct SimulatedExchange {
    pub id: Venue,
    pub oms_type: OmsType,
    pub account_type: AccountType,
    pub base_currency: Option<Currency>,
    /* private fields */
}

Fields
id: Venueoms_type: OmsTypeaccount_type: AccountTypebase_currency: Option<Currency>
Implementations

impl SimulatedExchange

pub fn new( venue: Venue, oms_type: OmsType, account_type: AccountType, starting_balances: Vec<Money>, base_currency: Option<Currency>, default_leverage: Decimal, leverages: HashMap<InstrumentId, Decimal>, modules: Vec<Box<dyn SimulationModule>>, cache: Rc<RefCell<Cache>>, clock: Rc<RefCell<dyn Clock>>, fill_model: FillModel, fee_model: FeeModelAny, book_type: BookType, latency_model: Option<LatencyModel>, frozen_account: Option<bool>, bar_execution: Option<bool>, reject_stop_orders: Option<bool>, support_gtd_orders: Option<bool>, support_contingent_orders: Option<bool>, use_position_ids: Option<bool>, use_random_ids: Option<bool>, use_reduce_only: Option<bool>, use_message_queue: Option<bool>, ) -> Result<Self>

Creates a new SimulatedExchange instance.

Errors

Returns an error if:

starting_balances is empty.

base_currency is Some but starting_balances contains multiple currencies.

pub fn register_client(&mut self, client: Rc<dyn ExecutionClient>)

pub fn set_fill_model(&mut self, fill_model: FillModel)

pub const fn set_latency_model(&mut self, latency_model: LatencyModel)

pub fn initialize_account(&mut self)

pub fn add_instrument(&mut self, instrument: InstrumentAny) -> Result<()>

Adds an instrument to the simulated exchange and initializes its matching engine.

Errors

Returns an error if the exchange account type is Cash and the instrument is a CryptoPerpetual or CryptoFuture.

Panics

Panics if the instrument cannot be added to the exchange.

pub fn best_bid_price(&self, instrument_id: InstrumentId) -> Option<Price>

pub fn best_ask_price(&self, instrument_id: InstrumentId) -> Option<Price>

pub fn get_book(&self, instrument_id: InstrumentId) -> Option<&OrderBook>

pub fn get_matching_engine( &self, instrument_id: &InstrumentId, ) -> Option<&OrderMatchingEngine>

pub const fn get_matching_engines( &self, ) -> &HashMap<InstrumentId, OrderMatchingEngine>

pub fn get_books(&self) -> HashMap<InstrumentId, OrderBook>

pub fn get_open_orders( &self, instrument_id: Option<InstrumentId>, ) -> Vec<PassiveOrderAny>

pub fn get_open_bid_orders( &self, instrument_id: Option<InstrumentId>, ) -> Vec<PassiveOrderAny>

pub fn get_open_ask_orders( &self, instrument_id: Option<InstrumentId>, ) -> Vec<PassiveOrderAny>

pub fn get_account(&self) -> Option<AccountAny>

Panics

Panics if retrieving the account from the execution client fails.

pub fn adjust_account(&mut self, adjustment: Money)

Panics

Panics if generating account state fails during adjustment.

pub fn send(&mut self, command: TradingCommand)

pub fn generate_inflight_command( &mut self, command: &TradingCommand, ) -> (UnixNanos, u32)

Panics

Panics if the command is invalid when generating inflight command.

pub fn process_order_book_delta(&mut self, delta: OrderBookDelta)

Panics

Panics if adding a missing instrument during delta processing fails.

pub fn process_order_book_deltas(&mut self, deltas: OrderBookDeltas)

Panics

Panics if adding a missing instrument during deltas processing fails.

pub fn process_quote_tick(&mut self, quote: &QuoteTick)

Panics

Panics if adding a missing instrument during quote tick processing fails.

pub fn process_trade_tick(&mut self, trade: &TradeTick)

Panics

Panics if adding a missing instrument during trade tick processing fails.

pub fn process_bar(&mut self, bar: Bar)

Panics

Panics if adding a missing instrument during bar processing fails.

pub fn process_instrument_status(&mut self, status: InstrumentStatus)

Panics

Panics if adding a missing instrument during instrument status processing fails.

pub fn process(&mut self, ts_now: UnixNanos)

Panics

Panics if popping an inflight command fails during processing.

pub fn reset(&mut self)

pub fn process_trading_command(&mut self, command: TradingCommand)

Panics

Panics if execution client is uninitialized when processing trading command.

pub fn generate_fresh_account_state(&self)

Panics

Panics if generating fresh account state fails.